Project Upload
First And Only Commit

+++ src/main/java/com/filiera/controller/AnimatoreController.java
@@ -4,6 +4,7 @@ package com.filiera.controller;
 import com.filiera.model.dto.EventoRequestDTO;
 
 import com.filiera.model.events.Evento;
+import com.filiera.model.events.Invito;
 import com.filiera.model.users.User;
 import com.filiera.services.AnimatoreService;
 import com.filiera.services.CurrentUserService;
@@ -36,10 +37,38 @@ public class AnimatoreController {
         return new ResponseEntity<>(newEvento, HttpStatus.CREATED);
     }
 
+    @PostMapping("/{eventId}/invite/{sellerId}")
+    public ResponseEntity<String> inviteSellerToEvent(
+            @PathVariable UUID eventId,
+            @PathVariable UUID sellerId
+    ) {
+        UUID organizerId = currentUserService.getCurrentUserId();
+
+        if (!animatoreService.checkPublisher(organizerId, eventId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("Non sei l'organizzatore di questo evento");
+        }
+
+
+        Invito invito = animatoreService.inviteSellerToEvent(eventId, sellerId);
+        return ResponseEntity.ok("Venditore invitato con successo all'evento!\nID invito :" + invito.getId());
+    }
+
+
     @DeleteMapping("/{eventId}")
-    public ResponseEntity<Void> deleteEvent(@PathVariable UUID eventId) {
+    public ResponseEntity<String> deleteEvent(@PathVariable UUID eventId) {
+
+        UUID organizerId = currentUserService.getCurrentUserId();
+
+        if (!animatoreService.checkPublisher(organizerId, eventId)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("Non sei l'organizzatore di questo evento");
+        }
+
         animatoreService.deleteEvent(eventId);
         // Return a 204 No Content status on successful deletion
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        return ResponseEntity.ok("Evento eliminato correttamente");
     }
+
+
 }
